add flag tests
